Moving RunnerManager class into manager module
